+++ adapters/Tapjoy/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Tapjoy Mediation Adapter for Google Mobile Ads SDK for iOS
 
+## Unreleased
+- Updated the default reward amount to 1.
+
 ## Version 11.10.1.0
 - Verified compatibility with Tapjoy SDK 11.10.1.