+++ adapters/Tapjoy/CHANGELOG.md
@@ -1,7 +1,12 @@
 # Tapjoy Mediation Adapter for Google Mobile Ads SDK for iOS
 
-## Unreleased
+## Version 11.11.0.0
+- Verified compatibility with Tapjoy SDK 11.11.0.
+
+## Version 11.10.2.0
 - Updated the default reward amount to 1.
+- Removed the support for `armv7s` architecture.
+- Verified compatibility with Tapjoy SDK 11.10.2.
 
 ## Version 11.10.1.0
 - Verified compatibility with Tapjoy SDK 11.10.1.